// Atspm/WebUI/src/api/identity/atspmAuthenticationApi.schemas.ts
// #region license
// Copyright 2026 Utah Departement of Transportation
// for WebUI - atspmAuthenticationApi.schemas.ts
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//http://www.apache.org/licenses/LICENSE-2.
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
// #endregion

/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * Atspm Authentication Api
 * OpenAPI spec version: v1
 */
export interface VerifyUserPasswordResetViewModel {
  /** @minLength 1 */
  password: string;
}

export interface VerifyResetTokenViewModel {
  /** @minLength 1 */
  token: string;
  /** @minLength 1 */
  username: string;
}

export interface VerifyConnectTokenViewModel {
  /** @minLength 1 */
  username: string;
}

export interface UserDTO {
  /** @nullable */
  agency?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  readonly fullName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  roles?: string[] | null;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  userName?: string | null;
}

export interface UpdateProfileViewModel {
  /**
   * @minLength 0
   * @maxLength 100
   */
  agency: string;
  /** @minLength 1 */
  email: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  firstName: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  lastName: string;
  /** @nullable */
  phoneNumber?: string | null;
}

export interface RegisterViewModel {
  /** @minLength 1 */
  agency: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 1 */
  password: string;
}

export interface ProblemDetails {
  /** @nullable */
  detail?: string | null;
  /** @nullable */
  instance?: string | null;
  /** @nullable */
  status?: number | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  type?: string | null;
  [key: string]: unknown;
}

export interface LoginViewModel {
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  password: string;
  /** @nullable */
  provider?: string | null;
  rememberMe?: boolean;
  /** @nullable */
  returnUrl?: string | null;
}

export interface ForgotPasswordViewModel {
  /** @minLength 1 */
  email: string;
}

export interface CreateRoleViewModel {
  /** @minLength 1 */
  roleName: string;
}

export interface ClaimsModel {
  /** @nullable */
  claims?: string[] | null;
}

export interface ClaimModel {
  /** @nullable */
  type?: string | null;
  /** @nullable */
  value?: string | null;
}

export interface ChangePasswordViewModel {
  /** @nullable */
  confirmPassword?: string | null;
  /** @minLength 1 */
  newPassword: string;
  /** @minLength 1 */
  resetToken: string;
}